Video-Storage-API/requirements.txt
fastapi
azure-storage-blob
python-multipart
aiofiles